refactor!: reorganize project structure
Didn't like how project was structured and decided to change.

BREAKING CHANGE: change import paths
