--- bababel/consumer/__init__.py
@@ -1,3 +1,0 @@
-from bababel.consumer.consumer import Consumer
-
-__all__ = Consumer
