+++ bababel/exceptions.py
@@ -5,11 +5,3 @@ class BababelError(Exception):
         self.message = message
         self.code = code
         self.details = details
-
-
-class TaskError(BababelError):
-    pass
-
-
-class MessageRouterException(BababelError):
-    pass
