--- bababel/message_router/__init__.py
@@ -1,3 +1,0 @@
-from bababel.message_router.message_router import MessageRouter
-
-__all__ = ['MessageRouter']
