--- bababel/publisher/__init__.py
@@ -1,3 +1,0 @@
-from bababel.publisher.rabbitmq_publisher import RabbitMQPublisher
-
-__all__ = ['RabbitMQPublisher']
