+++ bababel/routers/exceptions.py
@@ -1,0 +1,5 @@
+from bababel.exceptions import BababelError
+
+
+class MessageRouterException(BababelError):
+    pass
