--- bababel/task/__init__.py
@@ -1,3 +1,0 @@
-from bababel.task.task import Task
-
-__all__ = ['Task']
