+++ bababel/tasks/exceptions.py
@@ -1,0 +1,5 @@
+from bababel.exceptions import BababelError
+
+
+class TaskError(BababelError):
+    pass
